requirements.txt
PySide6
eventlet
flask
flask-cors
flask-socketio
python-socketio
websocket-client
opencv-python
numpy
requests
python-dotenv
aiohttp
onnxruntime
huggingface-hub>=0.19.0,<0.25.0
websockets
aiortc
gradio==4.29.0
matplotlib
psutil
cryptography
qrcode
pypng

# PTZ camera control
pytapo>=3.3.0,<4
onvif-zeep>=0.2.12

# Local services (Vision/LLM) runtime
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
prometheus-client>=0.19.0
# Keep compatible with gradio==4.29.0
aiofiles>=23.0.0,<24.0.0
pillow>=10.0.0,<11.0.0
transformers>=4.38.0,<5.0.0
accelerate>=0.24.1
sentencepiece>=0.1.99

# Inference (required for local vision; used by local LLM on-demand)
torch>=2.1.0
torchvision>=0.16.0